feat: add how-to guide for using LLMs and MCP in app plugins
This commit adds a new how-to guide for app plugin authors,
demonstrating how to use the `@grafana/llm` package to make
requests to LLMs, and how to make use of the Grafana MCP
server embedded into the LLM app.

I've combined these into one guide since the concepts are closely
related but I could split them out if that's preferred.